+++ src/util/featureconverter/clampedPrimitive.ts
@@ -1,6 +1,7 @@
 import {
   Cartesian3,
   Cartographic,
+  ClassificationPrimitive,
   HeightReference,
   Matrix4,
   Model,
@@ -37,19 +38,66 @@ function getUpdateHeightCallback(
   };
 }
 
+function getUpdateClassificationHeightCallback(
+  primitive: ClassificationPrimitive,
+  originalPosition: Cartesian3,
+): (cartographic: Cartographic) => void {
+  return function updatePrimitiveHeight(clampedPosition: Cartographic) {
+    Cartographic.toCartesian(
+      clampedPosition,
+      undefined,
+      scratchUpdateHeightCartesian,
+    );
+
+    Cartesian3.subtract(
+      scratchUpdateHeightCartesian,
+      originalPosition,
+      scratchUpdateHeightCartesian,
+    );
+
+    if (primitive.ready) {
+      // eslint-disable-next-line no-underscore-dangle
+      const innerPrimitive = primitive._primitive as Primitive;
+
+      const { modelMatrix } = innerPrimitive;
+      innerPrimitive.modelMatrix = Matrix4.setTranslation(
+        modelMatrix,
+        scratchUpdateHeightCartesian,
+        new Matrix4(),
+      );
+    } else {
+      // eslint-disable-next-line no-underscore-dangle
+      primitive._primitiveOptions!.modelMatrix = Matrix4.setTranslation(
+        Matrix4.IDENTITY,
+        scratchUpdateHeightCartesian,
+        new Matrix4(),
+      );
+    }
+  };
+}
+
 // eslint-disable-next-line import/prefer-default-export
 export function setupClampedPrimitive(
   scene: Scene,
-  primitive: Primitive | Model,
+  primitive: Primitive | Model | ClassificationPrimitive,
   origin: [number, number],
   heightReference: HeightReference,
 ): void {
   const destroy = primitive.destroy.bind(primitive);
   const originCartographic = mercatorToCartographic(origin);
-  const callback = getUpdateHeightCallback(
-    primitive,
-    Cartographic.toCartesian(originCartographic),
-  );
+  let callback;
+  if (primitive instanceof ClassificationPrimitive) {
+    callback = getUpdateClassificationHeightCallback(
+      primitive,
+      Cartographic.toCartesian(originCartographic),
+    );
+  } else {
+    callback = getUpdateHeightCallback(
+      primitive,
+      Cartographic.toCartesian(originCartographic),
+    );
+  }
+
   const callbackHandler = scene.updateHeight(
     originCartographic,
     callback,